Move tile producers to dedicated module
