+++ src/game/mod.rs
@@ -1,7 +1,8 @@
+pub mod tile;
+
 use core::{
     f32::consts::{FRAC_PI_2, PI},
     fmt::Debug,
-    mem::replace,
     ops::FnOnce,
 };
 use either::Either;
@@ -11,6 +12,7 @@ use microtile_engine::{
     geometry::tile::BasicTile,
 };
 use rtic_sync::channel::{ReceiveError, Receiver};
+use tile::TileProducer;
 
 #[must_use]
 pub enum Message {
@@ -22,82 +24,6 @@ pub enum Message {
 impl Message {
     pub fn acceleration(x: i16, z: i16) -> Self {
         Self::AccelerometerData { x, z }
-    }
-}
-
-pub trait TileProducer {
-    fn generate_tile(&mut self) -> BasicTile;
-}
-
-pub struct TileIterator<P> {
-    producer: P,
-}
-
-impl<P> TileIterator<P> {
-    #[must_use]
-    pub fn new(producer: P) -> Self {
-        Self { producer }
-    }
-}
-
-impl<P> Iterator for TileIterator<P>
-where
-    P: TileProducer,
-{
-    type Item = BasicTile;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        Some(self.producer.generate_tile())
-    }
-}
-
-pub struct ConstantProducer {
-    t: BasicTile,
-}
-
-impl ConstantProducer {
-    #[must_use]
-    pub fn new(t: BasicTile) -> Self {
-        Self { t }
-    }
-}
-
-impl TileProducer for ConstantProducer {
-    fn generate_tile(&mut self) -> BasicTile {
-        self.t.clone()
-    }
-}
-
-pub struct LoopingProducer {
-    t: BasicTile,
-}
-
-impl LoopingProducer {
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            t: BasicTile::Square,
-        }
-    }
-
-    fn advance(&mut self) -> BasicTile {
-        match self.t {
-            BasicTile::Square => replace(&mut self.t, BasicTile::Line),
-            BasicTile::Line => replace(&mut self.t, BasicTile::Diagonal),
-            BasicTile::Diagonal => replace(&mut self.t, BasicTile::Square),
-        }
-    }
-}
-
-impl Default for LoopingProducer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl TileProducer for LoopingProducer {
-    fn generate_tile(&mut self) -> BasicTile {
-        self.advance()
     }
 }
